Replaced float by double
Suggested in discussion thread at issue #300

+++ qucs-activefilter/mfbfilter.cpp
@@ -67,8 +67,8 @@ void MFBfilter::createLowPassSchematic(QString &s)
     for (int i=1; i<=N2ord; i++) {
         stage = Sections.at(i-1);
         QString suffix1, suffix2;
-        float C1 = autoscaleCapacitor(stage.C1,suffix1);
-        float C2 = autoscaleCapacitor(stage.C2,suffix2);
+        double  C1 = autoscaleCapacitor(stage.C1,suffix1);
+        double  C2 = autoscaleCapacitor(stage.C2,suffix2);
         s += QString("<GND * 1 %1 380 0 0 0 0>\n").arg(200+dx);
         s += QString("<GND * 1 %1 350 0 0 0 0>\n").arg(360+dx);
         s += QString("<OpAmp OP%1 1 %2 270 -26 -70 1 0 \"1e6\" 1 \"15 V\" 0>\n").arg(1+(i-1)*Nop1).arg(390+dx);
@@ -141,8 +141,8 @@ void MFBfilter::createHighPassSchematic(QString &s)
     for (int i=1; i<=N2ord; i++) {
         stage = Sections.at(i-1);
         QString suffix1, suffix2;
-        float C1 = autoscaleCapacitor(stage.C1,suffix1);
-        float C2 = autoscaleCapacitor(stage.C2,suffix2);
+        double  C1 = autoscaleCapacitor(stage.C1,suffix1);
+        double  C2 = autoscaleCapacitor(stage.C2,suffix2);
         s += QString("<GND * 1 %1 380 0 0 0 0>\n").arg(200+dx);
         s += QString("<GND * 1 %1 350 0 0 0 0>\n").arg(360+dx);
         s += QString("<OpAmp OP%1 1 %2 270 -26 -70 1 0 \"1e6\" 1 \"15 V\" 0>\n").arg(1+(i-1)*Nop1).arg(390+dx);
@@ -213,8 +213,8 @@ void MFBfilter::createBandPassSchematic(QString &s)
     for (int i=1; i<=Sections.count(); i++) {
         stage = Sections.at(i-1);
         QString suffix1, suffix2;
-        float C1 = autoscaleCapacitor(stage.C1,suffix1);
-        float C2 = autoscaleCapacitor(stage.C2,suffix2);
+        double  C1 = autoscaleCapacitor(stage.C1,suffix1);
+        double  C2 = autoscaleCapacitor(stage.C2,suffix2);
         s += QString("<GND * 1 %1 380 0 0 0 0>\n").arg(200+dx);
         s += QString("<GND * 1 %1 350 0 0 0 0>\n").arg(360+dx);
         s += QString("<OpAmp OP%1 1 %2 270 -26 -70 1 0 \"1e6\" 1 \"15 V\" 0>\n").arg(1+(i-1)*Nop1).arg(390+dx);
@@ -270,16 +270,16 @@ void MFBfilter::createBandStopSchematic(QString &s)
 
 void MFBfilter::calcHighPass()
 {
-    float R1,R2,C1,C2;
-    float Wc = 2*pi*Fc;
-    float Nst = order/2 + order%2;
-    float Kv1 = pow(Kv,1.0/Nst);
+    double  R1,R2,C1,C2;
+    double  Wc = 2*pi*Fc;
+    double  Nst = order/2 + order%2;
+    double  Kv1 = pow(Kv,1.0/Nst);
 
     for (int k=1; k <= order/2; k++) {
-        float re = Poles.at(k-1).real();
-        float im = Poles.at(k-1).imag();
-        float B = -2.0*re;
-        float C = re*re + im*im;
+        double  re = Poles.at(k-1).real();
+        double  im = Poles.at(k-1).imag();
+        double  B = -2.0*re;
+        double  C = re*re + im*im;
 
         C1 = 10.0/Fc;
         C2 = C1/Kv1;
@@ -305,16 +305,16 @@ void MFBfilter::calcHighPass()
 
 void MFBfilter::calcLowPass()
 {
-    float R1,R2,R3,C1,C2;
-    float Wc = 2*pi*Fc;
-    float Nst = order/2 + order%2;
-    float Kv1 = pow(Kv,1.0/Nst);
+    double  R1,R2,R3,C1,C2;
+    double  Wc = 2*pi*Fc;
+    double  Nst = order/2 + order%2;
+    double  Kv1 = pow(Kv,1.0/Nst);
 
     for (int k=1; k <= order/2; k++) {
-        float re = Poles.at(k-1).real();
-        float im = Poles.at(k-1).imag();
-        float B = -2.0*re;
-        float C = re*re + im*im;
+        double  re = Poles.at(k-1).real();
+        double  im = Poles.at(k-1).imag();
+        double  B = -2.0*re;
+        double  C = re*re + im*im;
 
         C2 = (10.0/Fc); // ballpark formula giving usually a good value
         // C1 maximum, can be smaller to have a convenient value
@@ -346,17 +346,17 @@ void MFBfilter::calcLowPass()
 
 void MFBfilter::calcBandPass()
 {
-    float W0 = 2*pi*F0;
-    float R1,R2,R3,C1,C2;
+    double  W0 = 2*pi*F0;
+    double  R1,R2,R3,C1,C2;
     //float rho = Kv/Q;
     //float gamma = 1.0;
     int cnt = 1; 
-    float Kv1 = pow(Kv,1.0/order);
+    double  Kv1 = pow(Kv,1.0/order);
 
     if (order==1) {  // Filter contains only 1 1st-order section
-        float rho = Kv1/Q;
-        float beta = 1.0/Q;
-        float gamma = 1.0;
+        double  rho = Kv1/Q;
+        double  beta = 1.0/Q;
+        double  gamma = 1.0;
 
         C1 = 10.0/F0;
         C2 = C1*(rho*beta-gamma)/gamma;
@@ -380,21 +380,21 @@ void MFBfilter::calcBandPass()
     }
 
     for (int k=1; k <= order/2; k++) {
-        float re = Poles.at(k-1).real();
-        float im = Poles.at(k-1).imag();
-        float B = -2.0*re;
-        float C = re*re + im*im;
+        double  re = Poles.at(k-1).real();
+        double  im = Poles.at(k-1).imag();
+        double  B = -2.0*re;
+        double  C = re*re + im*im;
 
-        float H = C + 4.0*Q*Q;
-        float E = (1.0/B)*sqrt(0.5*(H+sqrt(H*H-(4.0*B*B*Q*Q))));
-        float F = (B*E)/Q;
-        float D = 0.5*(F+sqrt(F*F-4.0));
+        double  H = C + 4.0*Q*Q;
+        double  E = (1.0/B)*sqrt(0.5*(H+sqrt(H*H-(4.0*B*B*Q*Q))));
+        double  F = (B*E)/Q;
+        double  D = 0.5*(F+sqrt(F*F-4.0));
 
         qDebug()<<D<<E<<Q;
 
-        float rho = Kv1*sqrt(C)/Q;
-        float beta = D/E;
-        float gamma = D*D;
+        double  rho = Kv1*sqrt(C)/Q;
+        double  beta = D/E;
+        double  gamma = D*D;
 
         C1 = 10.0/F0;
         C2 = C1*(rho*beta-gamma)/gamma;
@@ -445,14 +445,14 @@ void MFBfilter::calcBandPass()
 
     if (order%2 != 0) { // Need to implement first-order section
 
-        float R1,R2,R3,C1;
+        double  R1,R2,R3,C1;
 
         int k = order/2 + 1;
-        float re = Poles.at(k-1).real();
-        float C = -re;
-        float rho = Kv1*C/Q;
-        float beta = C/Q;
-        float gamma = 1.0;
+        double  re = Poles.at(k-1).real();
+        double  C = -re;
+        double  rho = Kv1*C/Q;
+        double  beta = C/Q;
+        double  gamma = 1.0;
 
         C1 = 10.0/F0;
         if (C2<0) C2=C1;
